+++ sdk/storage/Azure.Storage.Files.DataLake/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Release History
 
+## 12.3.0-preview.1 (Unreleased)
+
+
 ## 12.2.0 (2020-05)
 - Added DataLakeFileClient.Upload() overload that allows setting metadata, permissions and umask.
 - Fixed bug where PathClient.Rename() was not functioning correctly with SAS.